move `SiteAdminCell`, `DetailFieldCell` to prod. code
