+++ GitHubAPIApp/GitHubAPIApp/UI/UserDetailUI/UserDetailCell.swift
@@ -183,3 +183,77 @@ public class UserDetailCell: UITableViewCell {
     }
 
 }
+
+//MARK: - profile view
+class ProfileView: UIView {
+    private let avatarWidth: CGFloat = 160
+    
+    private let stackViewSpacing: CGFloat = 20.0
+    
+    private let nameLabelFontSize: CGFloat = 40.0
+    
+    private let biographyLabelFontSize: CGFloat = 20.0
+    
+    private lazy var stackView: UIStackView = {
+       let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(view)
+        
+        NSLayoutConstraint.activate([
+            view.leadingAnchor.constraint(equalTo: leadingAnchor),
+            view.trailingAnchor.constraint(equalTo: trailingAnchor),
+            view.topAnchor.constraint(equalTo: topAnchor),
+            view.bottomAnchor.constraint(equalTo: bottomAnchor),
+        ])
+        
+        view.axis = .vertical
+        view.distribution = .fill
+        view.alignment = .center
+        view.spacing = stackViewSpacing
+        
+        view.addArrangedSubview(avatarImageView)
+        view.addArrangedSubview(nameLabel)
+        view.addArrangedSubview(biographyLabel)
+        return view
+    }()
+    
+    public private(set) lazy var avatarImageView: UIImageView = {
+        let view = UIImageView()
+        NSLayoutConstraint.activate([
+            view.widthAnchor.constraint(equalToConstant: avatarWidth),
+            view.widthAnchor.constraint(equalTo: view.heightAnchor)
+        ])
+        
+        return view
+    }()
+    
+    public private(set) lazy var nameLabel: UILabel = {
+        let label = UILabel()
+        label.textAlignment = .center
+        label.font = UIFont.systemFont(ofSize: nameLabelFontSize)
+        return label
+    }()
+    
+    public private(set) lazy var biographyLabel: UILabel = {
+        let label = UILabel()
+        label.textAlignment = .center
+        label.font = UIFont.systemFont(ofSize: biographyLabelFontSize)
+        return label
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        commonInit()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        commonInit()
+    }
+    
+    private func commonInit() {
+        _ = stackView
+    }
+    
+}
+
